agents/agent2/requirements.txt
fastapi
uvicorn[standard]
httpx
pyautogui
requests
openai
python-dotenv
sqlalchemy
psycopg2-binary